Fix override of shared-readwrite mounting config
basehub defines the list `custom.singleuserAdmin.extraVolumeMounts` and
provides two entries to it.

If a hub re-specifies this list with a single new entry, it will end up
overriding entire list, making the two entries defined by default be
lost. This is a challenge with using helm values that are list, they
combine by replacing eachother rather than appending to each other.

Due to that, whenever we want to add a mount, we need to re-specify the
mounts that are in the list by default.

This commit impacts the following cluster, hubs:

- 2i2c-uk, lis
- 2i2c, climatematch
- leap, staging and prod