Add topic documentation on cloud billing concepts
Cloud billing is a vast area with many nuances. However, because of our usage
patterns, we only need to understand a small-ish subset of it. This needs to be
a shared understanding, so we all have a common base to reason from.

Fixes https://github.com/2i2c-org/infrastructure/issues/3760