Allow self-merging new hub PRs
As part of https://github.com/2i2c-org/meta/issues/897, we are
working heavily on *standardizing* what constitutes a new hub. The
new hub is first turned up with some standard options, and any non-standard
customizations are done after the fact.

Engineers can still request and wait for review if they want to, but
I think this will increase our velocity without undue danger. We can
always revert this if needed - I feel this is a safe, small thing
to try.

Ref https://github.com/2i2c-org/meta/issues/897