2i2c-aws-us and catalystproject-africa: disable shared-volume-reporter
Specifically this cluster with its many hubs runs out of available pods
per node on its core nodes, forcing two core nodes instead of one to be
used. That in turn increase costs for the communities on this cluster
with tens of percent.

Without these pods that reports metrics only used for a graphana
dashboard to conclude 100% of disk space is available on EFS without a
storage limit - we could together with other optimizations become able
to get down to using a single core node.