Support restricting profile options with github teams *only*
This gets rid of restricting profile options based on github *org*
membership, rather than *teams* membership. This is a complete no-op
for any of our communities, as it turns out nobody is using this feature!
Getting rid of this makes sure that new communities will not use this
feature, as it relies on internal implementation details of the
GitHub authenticator and may be hard to upstream when the time
comes. See https://github.com/2i2c-org/infrastructure/issues/4021
for upstreaming plan.