cryocloud: Simplify profile-list options
This was triggered by conversations around https://2i2c.freshdesk.com/a/tickets/1592,
which led to me helping Tasha look at cloud costs. In addition to large
use of home directories, we discovered that the other major cause was
people selecting the 'medium' option but with small sized instances - so
most of the instance stays unused, but it does cost a lot.

This PR:
- Switch to putting environment selection first, and resource
  selection as a second level dropdown
- Put most user options in the smaller r5.xlarge, and only the two
  large options in r5.4xlarge.

This is a replication of the options in the NASA VEDA hub.

She's reaching out to various folksk to help cleanup the home
directories.