Coalesce daskhub into basehub
I have timeboxed myself to 30min, and I have verified that this
gives us a fairly seamless migration pathway for existing installations
of daskhub (no re-install needed), and allows us to mix multiple
dependencies (primarily binderhub-service and dask-gateway) more
easily. It does this by taking the following two actions:

- Make dask-gateway a conditional dependency of basehub,
  rather than daskhub
- Move all remaining daskhub/values.yaml config into basehub,
  either under dask-gateway, or under an extraConfig that's enabled
  conditionally (via `custom.daskGateway.enabled`)

The end result here is that `daskhub` is now a stub that is
*completely empty*, and simply passes through to basehub! So basehub
is where *all* the action is, and `daskhub` is a simple backwards
compatible passthrough.

I've tested this on the openscapes staging hub and verified that
everything works as it should! I only had to make the following
three config additions:

1. Set `basehub.dask-gateway.enabled` to `true`
2. Set `basehub.jupyterhub.custom.daskGateway.enabled` to `true`.
3. Set `basehub.jupyterhub.singleuser.cloudMetadata.blockWithIptables` to
   `true`. We may not need this, but I didn't want to investigate that in
   this spike.

Remaining things to do to complete this migration:

- [ ] Remove *all* references to `daskhub` in our deployer code.
      If something needs to know if dask-gateway is enabled, it
      should check for `basehub.dask-gateway.enabled`
- [ ] Make appropriate config changes to all existing daskhub
      installations so this is a seamless transition.
- [ ] Document the fact that some of our existing hub config
      will have a `daskhub` moniker as a historical artifact. I
      think the additional minor cognitive overhead is totally
      worth the fact that we won't have to re-install any hubs!

Ref https://github.com/2i2c-org/infrastructure/issues/4087