2i2c, binderhub-ui-demo: remove fixme workaround with no effect
I think the reasons for this ending up could have been that a restart
led to a already running build pod was detected, and that it then
reported the same error because it had not been re-created with the
relevant credentials setup after the binderhub pod was restarted.

I really don't know, but removing this config and starting a fresh build
pod works.