terraform, gcp: rename budget resource as a workaround
Apparently terraform fails to migrate a budget with a fixed number to a
budget referencing the last month, but recreating the budget worked fine
though. By renaming this resource we accomplished that.

This is what was shown repeatedly after each terraform apply without
forcing the re-creation:

```
Terraform will perform the following actions:

  # google_billing_budget.budget[0] will be updated in-place
  ~ resource "google_billing_budget" "budget" {
        id              = "censored"
        name            = "censored"
        # (3 unchanged attributes hidden)

      ~ amount {
          ~ last_period_amount = false -> true

          - specified_amount {
              - currency_code = "USD" -> null
              - nanos         = 0 -> null
                # (1 unchanged attribute hidden)
            }
        }

        # (4 unchanged blocks hidden)
    }

Plan: 0 to add, 1 to change, 0 to destroy.
````