terraform, aws: support specifying bucket specific tags
The intent is to allow 2i2c:hub-name to be specified. I saw no option
to easily and robustly add the 2i2c:hub-name tag automatically by using
other config or variables. The `hub_cloud_permissions` variable included
details about hub name, but I don't think we could be certain all
buckets would be tied to a hub name through that, so it seemed
problematic to try to use that information. With that in mind, I settled
for explicitly configuring any tags.