aws docs: add back now conditional step on adding cluster permissions
I now understand this better, and it turns out its still needed in k8s
clusters created for accounts where we aren't using 2i2c's AWS SSO as
then we don't access AWS with a single role but as separate users. This
is because the role or user creating the cluster gets access, where we
use the same role when working with 2i2c's AWS SSO credentials but
different users when working against other AWS accounts.

This reverts commit 0d4aa98ddeaa7907d79d5389df0a46366d878f8f.