Stop using `--wait` with helm deployments
While working on https://github.com/2i2c-org/infrastructure/pull/5809,
I finally got frustrated enough with our usage of `--wait` that I've
ripped it out.

The major issue with `--wait` is that if the deployment fails (because
of a typo in config for example), you *must* wait for the 5min timeout
or helm can get into an inconsistent state. This sucks when prototyping.

I stole the code we use in the mybinder.org deployment repo
instead (https://github.com/jupyterhub/mybinder.org-deploy/blob/main/deploy.py#L273).

It also seems faster, but the most important capability here is
that it's cancellable.