Pass ext-vars to jsonnet rather than reading our cluster.yaml files
I want to *eliminate* extra steps we were doing for jupyterhub-home-nfs
migration, and use that as a way to make our overall deployment
process more robust. Humans make mistakes (which is ok), so let's only
require humans to do things that require human judgement.

https://github.com/2i2c-org/infrastructure/pull/5939 eliminated the
need to copy paste config for enabling alerts, and we also then got
alerts for other cloud providers 'for free'.

This PR aims to remove the copy pasting for setting resources
on jupyterhub-home-nfs for each hub we deploy it to. We don't want
to set any limits for staging hubs, so we currently copy paste. This
approach allows us to have a single jsonnet file that can set the
limits for all non-staging hubs more easily.

This PR removes the feature from #5939 that allows jsonnet files to
read .yaml files from the directory of the hub they were deploying to.
This didn't allow us to differentiate between staging and production
hubs anyway, and also tied in knowledge about the structure of the
2i2c specific cluster.yaml files in jsonnet, which may lead to potential
right to replicate issues.

Instead, this PR sets a cluster name and (optionally) hub name as external
variables that can be read from jsonnet. This allows us to conditionally
make changes for staging hubs in a future PR.