Add alert for when hub db directory is about to be full
After merging https://github.com/2i2c-org/infrastructure/pull/6050,
I noticed that the utoronto staging hub was consistently failing
the health check. I investigated, and noticed that the hub disk
db is full! This is more prevalant on Azure because we also store
hub logs in there.

This could have also totally happened to a prod hub there, and would
have caused an outage - and it actually has in the past. We would like outages
like this to be prevented, by having alerts show up before the actual issue.

This PR does the following:

1. Add an alert for the hub db directory approaching fullness
2. Change the existing home directory alert to use the same metrics as
   hub db directory alert. This removes dependency on the extra shared-dir
   node exporter we run for alerts!
3. Use a jsonnet function to generate the alert rule, so we can more easily
   generate disk dir fullness alerts without having to duplicate the config.
4. Change the set of labels added to the alert so only actually helpful
   labels are set.