Alert when prometheus disk is running out of space
We were supposed to be *manually* checking prometheus disk
space every three months. However, this does not actually
happen in practice (see https://github.com/2i2c-org/infrastructure/issues/5830
sitting there for 2 months), and most importantly, *should not*
happen - this is something for a machine to do, not humans.

This was more difficult to do before because we did not have the
alerting infrastructure. We do now, so let's make use of that.

Fixes https://github.com/2i2c-org/infrastructure/issues/5830,
and stops that from ever happening