GHG: tune the CPU request for SSIM workshop
We tested spinning up multiple pods for the SSIM workshop. It always created a node per pod because the CPU availability was insufficient.

I'm decreasing the CPU request here a bit to test if that helps put 2 user pods in a single node for the workshop.

Changing the option for the non-workshop environment too just to help us test without being part of the workshop group.