Put contents of repo image is built from into /srv/repo
By default repo2docker puts the contents of the repo the image is
built from in /home/jovyan. This is overwritten by persistent home
directory storage when using a persistent hub, so users can not
access it.

This also causes problems when there's a `start` script in the repo. That is set to be the entrypoint,
and if that's not accessible during server startup, startup fails. And the user gets a mysterious error
message, as kubespawner doesn't actually give you server logs.

The solution in repo2docker is to put the contents in a different
place. We have been putting it `/srv/repo` as part of our
user image template with repo2docker-action
(https://github.com/2i2c-org/hub-user-image-template/blob/e0a2578e66825134accccbd34b36ca2802e29079/.github/workflows/build.yaml#L44).

This PR puts out the same behavior for all our dynamic imagebuilding
hubs.

Ref https://github.com/2i2c-org/infrastructure/issues/5499
Fixes https://github.com/2i2c-org/infrastructure/issues/9131