Add alert for server failing to start
This replaces the unschedulable pod alert, as this is more
accurate (hopefully). We had an UBC EOAS outage today that was not
caught by the unschedulable pod alert until much later.

I removed the unschedulable pod alert as it was added for the
same reason this one is. And I don't want us to have alert
fatigue