Setup cluster for oceanhackweek with new jsonnet functions
Our usage and understanding of jsonnet has grown with our team
in the last few years, and allows us to confidently use more
jsonnet features to simplify our config. Instead of treating our
jsonnet files as *text* (and templating them for reuse), we can treat them
as *code* (and use functions for reuse). This is very powerful,
and leads to simpler code overall that is easier to maintain and
make mass changes to.

Factors to consider in this design:

1. It *must* still admit customizations on a per cluster basis,
   particularly for version upgrades and similar as those will
   likely be done on a rolling basis.
2. It shouldn't stray too far from our existing config, so we
   don't end up with too many different variants. We can do
   some unification during our next migration
3. It should treat jsonnet as a programming language, and thus use
   various programming language constructs liberally. Since the
   language is declarative and our output is still a static YAML
   file, without any real side effects, I'm not super concerned about
   right to replicate concerns

This fundamentally addresses https://github.com/2i2c-org/infrastructure/issues/6632

When this PR is complete, it should set up the cluster for
the oceanhackweek-es hub: https://github.com/2i2c-org/infrastructure/issues/6700