Add sharedservices cluster with federated prometheus
Solving https://github.com/2i2c-org/meta/issues/2815
requires we have a centralized place to store long term metrics
for use by BD. This is a specific class of a problem we have -
that we have some centralized things we need to run that have
to be independent of each community's cluster. The centralized
grafana is a good example.

In the past, we have put that in our shared GCP cluster. However,
that is going away. And we have avoided running some services
because we don't have a shared space to run that from.

This PR:

- Sets up a sharedservices cluster to run such shared services
- Runs a prometheus server that is federating active user data
  only from all our prometheuses via prometheus [federation](https://prometheus.io/docs/prometheus/latest/federation/)
- Automatically sets up prometheus config by reading our repo,
  same as we do for uptime checks.
- Use terraform to set up the cluster *and* run helm, as we will
  (by definition) only have one of these

We already have uptime checks set up via stackdriver, and that's
a shared service. I re-use the same terraform here, and will rename
it shortly.

Prometheus federation already works well and is collecting data!